unified sessione-attiva nessuna-sessione conferma
